refactor(web): split monolithic server.go by domain
Move API handlers out of the 4k-line server.go into focused package
files (chat, sessions, models, mcp, providers, …) and append PTY/WS/SSH
handlers to existing modules. Also fix leftover composables path
references after the Vue → React migration.
