+++ internal/web/files.go
@@ -1,0 +1,374 @@
+package web
+
+import (
+	"bufio"
+	"context"
+	"encoding/json"
+	"io"
+	"net/http"
+	"os"
+	"os/exec"
+	"path/filepath"
+	"strings"
+
+	utils "github.com/cnjack/jcode/internal/util"
+)
+
+// withinWorkspace reports whether abs is the workspace root or strictly inside
+// it. Uses filepath.Rel rather than strings.HasPrefix so a sibling like /repo2
+// can't escape /repo, and an empty root rejects everything.
+func withinWorkspace(root, abs string) bool {
+	if root == "" {
+		return false
+	}
+	rel, err := filepath.Rel(root, abs)
+	if err != nil {
+		return false
+	}
+	return rel == "." || (rel != ".." && !strings.HasPrefix(rel, ".."+string(filepath.Separator)))
+}
+
+func (s *Server) handleListFiles(w http.ResponseWriter, r *http.Request) {
+	pwd := s.activePwd()
+	dir := r.URL.Query().Get("path")
+	if dir == "" {
+		dir = pwd
+	} else if !filepath.IsAbs(dir) {
+		dir = filepath.Join(pwd, dir)
+	}
+
+	// Prevent path traversal / sibling escape.
+	abs := filepath.Clean(dir)
+	if !withinWorkspace(pwd, abs) {
+		writeJSON(w, http.StatusBadRequest, map[string]string{"error": "invalid path"})
+		return
+	}
+
+	entries, err := os.ReadDir(abs)
+	if err != nil {
+		writeJSON(w, http.StatusInternalServerError, map[string]string{"error": err.Error()})
+		return
+	}
+
+	type fileItem struct {
+		Name  string `json:"name"`
+		IsDir bool   `json:"is_dir"`
+		Size  int64  `json:"size"`
+	}
+
+	items := make([]fileItem, 0, len(entries))
+	for _, e := range entries {
+		info, _ := e.Info()
+		size := int64(0)
+		if info != nil {
+			size = info.Size()
+		}
+		items = append(items, fileItem{
+			Name:  e.Name(),
+			IsDir: e.IsDir(),
+			Size:  size,
+		})
+	}
+	writeJSON(w, http.StatusOK, items)
+}
+
+func (s *Server) handleReadFile(w http.ResponseWriter, r *http.Request) {
+	path := r.URL.Query().Get("path")
+	if path == "" {
+		writeJSON(w, http.StatusBadRequest, map[string]string{"error": "path is required"})
+		return
+	}
+
+	pwd := s.activePwd()
+	abs := path
+	if !filepath.IsAbs(abs) {
+		abs = filepath.Join(pwd, abs)
+	}
+
+	// Prevent path traversal / sibling escape.
+	abs = filepath.Clean(abs)
+	if !withinWorkspace(pwd, abs) {
+		writeJSON(w, http.StatusBadRequest, map[string]string{"error": "path outside workspace"})
+		return
+	}
+
+	content, err := os.ReadFile(abs)
+	if err != nil {
+		writeJSON(w, http.StatusNotFound, map[string]string{"error": err.Error()})
+		return
+	}
+
+	// Limit file size to 1MB.
+	if len(content) > 1<<20 {
+		writeJSON(w, http.StatusRequestEntityTooLarge, map[string]string{
+			"error": "file too large (>1MB)",
+		})
+		return
+	}
+
+	writeJSON(w, http.StatusOK, map[string]string{
+		"path":    abs,
+		"content": string(content),
+	})
+}
+
+func (s *Server) handleExec(w http.ResponseWriter, r *http.Request) {
+	var req struct {
+		Command string `json:"command"`
+	}
+	if err := json.NewDecoder(io.LimitReader(r.Body, 1<<16)).Decode(&req); err != nil {
+		writeJSON(w, http.StatusBadRequest, map[string]string{"error": "invalid request"})
+		return
+	}
+	if strings.TrimSpace(req.Command) == "" {
+		writeJSON(w, http.StatusBadRequest, map[string]string{"error": "command is required"})
+		return
+	}
+
+	ctx, cancel := context.WithTimeout(s.rootCtx(), 30*1e9) // 30 seconds
+	defer cancel()
+
+	cmd := exec.CommandContext(ctx, "sh", "-c", req.Command)
+	cmd.Dir = s.activePwd()
+
+	output, err := cmd.CombinedOutput()
+	exitCode := 0
+	if err != nil {
+		if exitErr, ok := err.(*exec.ExitError); ok {
+			exitCode = exitErr.ExitCode()
+		} else {
+			writeJSON(w, http.StatusInternalServerError, map[string]string{"error": err.Error()})
+			return
+		}
+	}
+
+	// Truncate output to 256KB
+	out := string(output)
+	if len(out) > 256*1024 {
+		out = out[:256*1024] + "\n... (truncated)"
+	}
+
+	writeJSON(w, http.StatusOK, map[string]any{
+		"output":    out,
+		"exit_code": exitCode,
+	})
+}
+
+func (s *Server) handleDiff(w http.ResponseWriter, r *http.Request) {
+	mode := r.URL.Query().Get("mode")
+	if mode == "" {
+		mode = "working"
+	}
+
+	// "session" mode: diff between snapshot taken at agent run start and current state.
+	if mode == "session" {
+		s.handleSessionDiff(w, r)
+		return
+	}
+
+	var args []string
+	switch mode {
+	case "staged":
+		args = []string{"diff", "--cached", "--no-color"}
+	case "branch":
+		args = []string{"diff", "HEAD~1", "--no-color"}
+	default: // "working"
+		args = []string{"diff", "--no-color"}
+	}
+
+	cmd := exec.CommandContext(s.rootCtx(), "git", args...)
+	cmd.Dir = s.activePwd()
+	cmd.Env = utils.ScrubbedGitEnv()
+	output, _ := cmd.CombinedOutput()
+
+	// Parse diff into structured entries
+	type diffEntry struct {
+		File      string `json:"file"`
+		Patch     string `json:"patch"`
+		Additions int    `json:"additions"`
+		Deletions int    `json:"deletions"`
+		Status    string `json:"status"` // "M", "A", "D"
+	}
+
+	var entries []diffEntry
+	rawDiff := string(output)
+
+	// Also get changed file list for status
+	statCmd := exec.CommandContext(s.rootCtx(), "git", "diff", "--stat", "--no-color")
+	statCmd.Env = utils.ScrubbedGitEnv()
+	switch mode {
+	case "staged":
+		statCmd = exec.CommandContext(s.rootCtx(), "git", "diff", "--cached", "--stat", "--no-color")
+		statCmd.Env = utils.ScrubbedGitEnv()
+	case "branch":
+		statCmd = exec.CommandContext(s.rootCtx(), "git", "diff", "HEAD~1", "--stat", "--no-color")
+		statCmd.Env = utils.ScrubbedGitEnv()
+	}
+	statCmd.Dir = s.activePwd()
+	_, _ = statCmd.CombinedOutput()
+
+	// Parse unified diff into per-file entries
+	sections := splitDiffByFile(rawDiff)
+	for _, sec := range sections {
+		adds, dels := countDiffLines(sec.patch)
+		entries = append(entries, diffEntry{
+			File:      sec.file,
+			Patch:     sec.patch,
+			Additions: adds,
+			Deletions: dels,
+			Status:    sec.status,
+		})
+	}
+
+	if entries == nil {
+		entries = []diffEntry{}
+	}
+
+	writeJSON(w, http.StatusOK, map[string]any{
+		"mode":    mode,
+		"entries": entries,
+	})
+}
+
+type diffSection struct {
+	file   string
+	patch  string
+	status string
+}
+
+func splitDiffByFile(raw string) []diffSection {
+	var sections []diffSection
+	lines := strings.Split(raw, "\n")
+	var current *diffSection
+	var patchLines []string
+
+	for _, line := range lines {
+		if strings.HasPrefix(line, "diff --git ") {
+			// Flush previous
+			if current != nil {
+				current.patch = strings.Join(patchLines, "\n")
+				sections = append(sections, *current)
+			}
+			// Parse file name from "diff --git a/foo b/foo"
+			parts := strings.SplitN(line, " b/", 2)
+			file := ""
+			if len(parts) == 2 {
+				file = parts[1]
+			}
+			current = &diffSection{file: file, status: "M"}
+			patchLines = []string{line}
+		} else if current != nil {
+			patchLines = append(patchLines, line)
+			if strings.HasPrefix(line, "new file") {
+				current.status = "A"
+			} else if strings.HasPrefix(line, "deleted file") {
+				current.status = "D"
+			}
+		}
+	}
+	if current != nil {
+		current.patch = strings.Join(patchLines, "\n")
+		sections = append(sections, *current)
+	}
+	return sections
+}
+
+func countDiffLines(patch string) (adds, dels int) {
+	scanner := bufio.NewScanner(strings.NewReader(patch))
+	for scanner.Scan() {
+		line := scanner.Text()
+		if strings.HasPrefix(line, "+") && !strings.HasPrefix(line, "+++") {
+			adds++
+		} else if strings.HasPrefix(line, "-") && !strings.HasPrefix(line, "---") {
+			dels++
+		}
+	}
+	return
+}
+
+// takeSessionSnapshot records the current git working tree state
+// so that session-scoped diffs can be computed later.
+func (s *Server) takeSessionSnapshot(eng *Engine) {
+	if eng == nil {
+		return
+	}
+	// Use "git stash create" to get a tree-ish of the current state without
+	// actually stashing. If there are no changes, use HEAD.
+	cmd := exec.CommandContext(s.rootCtx(), "git", "stash", "create")
+	cmd.Dir = eng.pwd
+	cmd.Env = utils.ScrubbedGitEnv()
+	out, err := cmd.Output()
+	snapshot := strings.TrimSpace(string(out))
+	if err != nil || snapshot == "" {
+		// No local changes — use HEAD as baseline
+		cmd2 := exec.CommandContext(s.rootCtx(), "git", "rev-parse", "HEAD")
+		cmd2.Dir = eng.pwd
+		cmd2.Env = utils.ScrubbedGitEnv()
+		out2, _ := cmd2.Output()
+		snapshot = strings.TrimSpace(string(out2))
+	}
+	eng.emu.Lock()
+	eng.sessionSnapshot = snapshot
+	eng.emu.Unlock()
+}
+
+// handleSessionDiff computes the diff between the session start snapshot and current state.
+func (s *Server) handleSessionDiff(w http.ResponseWriter, _ *http.Request) {
+	// Capture the active engine ONCE so the snapshot and the working dir come
+	// from the same task's repo even if the active engine is swapped between the
+	// two reads (otherwise we could diff engine A's snapshot against engine B's
+	// tree). eng.pwd is immutable after creation, so reading it bare is safe.
+	eng := s.activeEngine()
+	snapshot := ""
+	pwd := ""
+	if eng != nil {
+		eng.emu.Lock()
+		snapshot = eng.sessionSnapshot
+		eng.emu.Unlock()
+		pwd = eng.pwd
+	}
+
+	type diffEntry struct {
+		File      string `json:"file"`
+		Patch     string `json:"patch"`
+		Additions int    `json:"additions"`
+		Deletions int    `json:"deletions"`
+		Status    string `json:"status"`
+	}
+
+	if snapshot == "" {
+		writeJSON(w, http.StatusOK, map[string]any{
+			"mode":    "session",
+			"entries": []diffEntry{},
+		})
+		return
+	}
+
+	// Diff from snapshot to current working tree
+	cmd := exec.CommandContext(s.rootCtx(), "git", "diff", snapshot, "--no-color")
+	cmd.Dir = pwd
+	cmd.Env = utils.ScrubbedGitEnv()
+	output, _ := cmd.CombinedOutput()
+
+	var entries []diffEntry
+	sections := splitDiffByFile(string(output))
+	for _, sec := range sections {
+		adds, dels := countDiffLines(sec.patch)
+		entries = append(entries, diffEntry{
+			File:      sec.file,
+			Patch:     sec.patch,
+			Additions: adds,
+			Deletions: dels,
+			Status:    sec.status,
+		})
+	}
+
+	if entries == nil {
+		entries = []diffEntry{}
+	}
+
+	writeJSON(w, http.StatusOK, map[string]any{
+		"mode":    "session",
+		"entries": entries,
+	})
+}
